Created self adjusting sensor threshold
The sensor threshold now updates itself by constantly checking if the
current reading is lower than the minimum value recorded or higher than
the max value recorded.  If it is, the min and max values are reset and
the threshold is set to the average of the two readings.  The threshold
starts as 100 based on tests using this setup.